Cleaning up Files Making things safer Observing the text in the textview cleaning up TextView API
