+++ ComfyEditor/Views/ComfyEditor/Components/ComfyEditorFrameView.swift
